Add integer/number enum support and improve canParse
- Extend IrEnum with valueKind to support int and double enums, not
  just string enums. Enum emitter generates typed value fields,
  fromJson/toJson with correct types, and deduplicates switch cases.
- canParse now checks for at least one known property key when a model
  has no required fields, instead of returning true unconditionally.

+++ snapshots/public/cloudflare/lib/models/tls_certificates_and_hostnames_certificates.dart
@@ -23,7 +23,7 @@ factory TlsCertificatesAndHostnamesCertificates.fromJson(Map<String, dynamic> js
   hostnames: (json['hostnames'] as List<dynamic>).map((e) => e as String).toList(),
   id: json['id'] != null ? TlsCertificatesAndHostnamesIdentifier.fromJson(json['id'] as String) : null,
   requestType: TlsCertificatesAndHostnamesRequestType.fromJson(json['request_type'] as String),
-  requestedValidity: TlsCertificatesAndHostnamesRequestedValidity.fromJson(json['requested_validity'] as num),
+  requestedValidity: TlsCertificatesAndHostnamesRequestedValidity.fromJson((json['requested_validity'] as num).toDouble()),
 ); }
 
 final TlsCertificatesAndHostnamesCertificatesComponentsSchemasCertificate? certificate;
